unzip downloaded files before feeding into EN on iOS
