+++ src/bridge/ExposureNotification.ts
@@ -6,71 +6,15 @@
 
 import {NativeModules} from 'react-native';
 
-export enum RiskLevel {
-  Invalid = 0,
-  Lowest = 1,
-  Low = 2,
-  LowMedium = 3,
-  Medium = 4,
-  MediumHigh = 5,
-  High = 6,
-  VeryHigh = 7,
-  Highest = 8,
-}
+import ExposureNotificationAdapter from './ExposureNotificationAdapter';
 
-export enum Status {
-  Unknown = 'unknown',
-  Active = 'active',
-  Disabled = 'disabled',
-  BluetoothOff = 'bluetooth_off',
-  Restricted = 'restricted',
-}
-
-export interface TemporaryExposureKey {
-  keyData: string;
-  rollingStartNumber: number;
-  rollingPeriod: number;
-  transmissionRiskLevel: RiskLevel;
-}
-
-export interface ExposureSummary {
-  daysSinceLastExposure: number;
-  matchedKeyCount: number;
-  maximumRiskScore: number;
-}
-
-export interface ExposureConfiguration {
-  metadata?: object;
-  minimumRiskScore: number;
-  attenuationLevelValues: number[];
-  attenuationWeight: number;
-  daysSinceLastExposureLevelValues: number[];
-  daysSinceLastExposureWeight: number;
-  durationLevelValues: number[];
-  durationWeight: number;
-  transmissionRiskLevelValues: number[];
-  transmissionRiskWeight: number;
-}
-
-export interface ExposureInformation {
-  dateMillisSinceEpoch: number;
-  durationMinutes: number;
-  attenuationValue: number;
-  transmissionRiskLevel: RiskLevel;
-  totalRiskScore: number;
-}
-
-export interface ExposureNotification {
-  start(): Promise<void>;
-  stop(): Promise<void>;
-  resetAllData(): Promise<void>;
-
-  getStatus(): Promise<Status>;
-
-  getTemporaryExposureKeyHistory(): Promise<TemporaryExposureKey[]>;
-
-  detectExposure(configuration: ExposureConfiguration, diagnosisKeysURLs: string[]): Promise<ExposureSummary>;
-  getExposureInformation(summary: ExposureSummary): Promise<ExposureInformation[]>;
-}
-
-export default NativeModules.ExposureNotification as ExposureNotification;
+export {
+  RiskLevel,
+  Status,
+  TemporaryExposureKey,
+  ExposureSummary,
+  ExposureConfiguration,
+  ExposureInformation,
+  ExposureNotification,
+} from './ExposureNotificationAPI';
+export default ExposureNotificationAdapter(NativeModules.ExposureNotification);
