+++ src/bridge/ExposureNotificationAdapter.d.ts
@@ -1,0 +1,5 @@
+import {ExposureNotification} from './ExposureNotificationAPI';
+
+declare function ExposureNotificationAdapter(exposureNotificationAPI: ExposureNotification): ExposureNotification;
+
+export default ExposureNotificationAdapter;
